+++ test/Sentry.Maui.Device.TestApp/Startup.cs
@@ -1,4 +1,6 @@
+#if VISUAL_RUNNER
 using DeviceRunners.VisualRunners;
+#endif
 using DeviceRunners.XHarness;
 using Microsoft.Maui.LifecycleEvents;
 
@@ -8,8 +10,16 @@ public static class MauiProgram
 {
     public static MauiApp CreateMauiApp()
     {
-        var appBuilder = MauiApp
-            .CreateBuilder()
+        var assemblies = new List<Assembly>(
+        [
+            typeof(Sentry.Tests.SentrySdkTests).Assembly,
+            typeof(Sentry.Extensions.Logging.Tests.LogLevelExtensionsTests).Assembly,
+            typeof(Sentry.Maui.Tests.SentryMauiOptionsTests).Assembly,
+#if ANDROID
+           typeof(Sentry.Android.AssemblyReader.Tests.AndroidAssemblyReaderTests).Assembly,
+#endif
+        ]);
+        var appBuilder = MauiApp.CreateBuilder()
             .ConfigureLifecycleEvents(life =>
             {
 #if __ANDROID__
@@ -21,29 +31,17 @@ public static class MauiProgram
             })
             .UseXHarnessTestRunner(conf =>
             {
-                conf.AddTestAssemblies([
-                    typeof(Sentry.Tests.SentrySdkTests).Assembly,
-                    typeof(Sentry.Extensions.Logging.Tests.LogLevelExtensionsTests).Assembly,
-                    typeof(Sentry.Maui.Tests.SentryMauiOptionsTests).Assembly,
-#if ANDROID
-                    typeof(Sentry.Android.AssemblyReader.Tests.AndroidAssemblyReaderTests).Assembly,
-#endif
-                ]);
+                conf.AddTestAssemblies(assemblies);
                 conf.AddXunit();
             });
 
-        //             .UseVisualTestRunner(conf =>
-        //             {
-        //                 conf.AddTestAssemblies([
-        //                     typeof(Sentry.Tests.SentrySdkTests).Assembly,
-        //                     typeof(Sentry.Extensions.Logging.Tests.LogLevelExtensionsTests).Assembly,
-        //                     typeof(Sentry.Maui.Tests.SentryMauiOptionsTests).Assembly,
-        // #if ANDROID
-        //                     typeof(Sentry.Android.AssemblyReader.Tests.AndroidAssemblyReaderTests).Assembly,
-        // #endif
-        //                 ]);
-        //                 conf.AddXunit();
-        //             });
+#if VISUAL_RUNNER
+        appBuilder.UseVisualTestRunner(conf =>
+        {
+            conf.AddTestAssemblies(assemblies);
+            conf.AddXunit();
+        });
+#endif
 
         return appBuilder.Build();
     }